% Should be automatically generated using roxygen2

\name{rvm}
\alias{rvm}
\alias{rvm}
\alias{rvm}
\alias{summary.rvm}
\alias{print.summary.rvm}
\alias{print.rvm}
\title{Relevance Vector Machines}
\description{
\code{rvm} is used to train a relevance vector machine. It can be used to carry
out general regression and classification. A formula interface is provided.
}
\references{
  \itemize{
    \item
      Tipping:\cr
      \emph{Sparse Bayesian Learning and the Relevance Vector Machine}\cr
      \url{http://www.jmlr.org/papers/volume1/tipping01a/tipping01a.pdf}
    \item
      Tipping, Faul:\cr
      \emph{Fast Marginal Likelihood Maximisation for Sparse Bayesian Models}\cr
      \url{http://www.miketipping.com/papers/met-fastsbl.pdf}
  }
}
